requirements.txt
flet>=0.23.0
rembg>=2.0.60
pillow>=11.0.0
onnxruntime>=1.20.1